+++ src/compas/files/gltf/data_classes.py
@@ -3,7 +3,61 @@ from __future__ import division
 from __future__ import absolute_import
 
 
-class SamplerData(object):
+class AlphaMode(object):
+    BLEND = "BLEND"
+    MASK = "MASK"
+    OPAQUE = "OPAQUE"
+
+
+class MineType(object):
+    JPEG = "image/jpeg"
+    PNG = "image/png"
+
+
+# I changed the name of this so as not to collide with compas.Base
+class BaseGLTFDataClass(object):
+
+    IS_BASE_GLTF_DATA = True  # only needed for ipy in `GLTFContent.check_extensions_texture_recursively`
+
+    def __init__(self, extras=None, extensions=None):
+        self.extras = extras
+        self.extensions = extensions
+
+    def add_extension(self, extension):
+        if not self.extensions:
+            self.extensions = {}
+        self.extensions.update({extension.key: extension})
+
+    def extensions_to_data(self, **kwargs):
+        if not self.extensions:
+            return None
+        return {key: value.to_data(**kwargs) if hasattr(value, "to_data") else value for key, value in self.extensions.items()}
+
+    @classmethod
+    def extensions_from_data(cls, data):
+        # i hate hate hate this local import, but i don't see a good way around it
+        from compas.files.gltf.extensions import SUPPORTED_EXTENSIONS
+
+        if not data:
+            return None
+        extensions = {}
+        for key, value_data in data.items():
+            if key in SUPPORTED_EXTENSIONS:
+                extensions[key] = SUPPORTED_EXTENSIONS[key].from_data(value_data)
+            else:
+                extensions[key] = value_data
+        return extensions
+
+    def to_data(self, *args, **kwargs):
+        dct = {}
+        if self.extras is not None:
+            dct["extras"] = self.extras
+        if self.extensions is not None:
+            dct["extensions"] = self.extensions_to_data()
+        return dct
+
+
+class SamplerData(BaseGLTFDataClass):
     def __init__(
         self,
         mag_filter=None,
@@ -14,30 +68,29 @@ class SamplerData(object):
         extras=None,
         extensions=None,
     ):
+        super(SamplerData, self).__init__(extras, extensions)
         self.mag_filter = mag_filter
         self.min_filter = min_filter
         self.wrap_s = wrap_s
         self.wrap_t = wrap_t
         self.name = name
-        self.extras = extras
-        self.extensions = extensions
 
     def to_data(self):
         sampler_dict = {}
         if self.mag_filter is not None:
-            sampler_dict['magFilter'] = self.mag_filter
+            sampler_dict["magFilter"] = self.mag_filter
         if self.min_filter is not None:
-            sampler_dict['minFilter'] = self.min_filter
+            sampler_dict["minFilter"] = self.min_filter
         if self.wrap_s is not None:
-            sampler_dict['wrapS'] = self.wrap_s
+            sampler_dict["wrapS"] = self.wrap_s
         if self.wrap_t is not None:
-            sampler_dict['wrapT'] = self.wrap_t
+            sampler_dict["wrapT"] = self.wrap_t
         if self.name is not None:
-            sampler_dict['name'] = self.name
+            sampler_dict["name"] = self.name
         if self.extras is not None:
-            sampler_dict['extras'] = self.extras
+            sampler_dict["extras"] = self.extras
         if self.extensions is not None:
-            sampler_dict['extensions'] = self.extensions
+            sampler_dict["extensions"] = self.extensions_to_data()
         return sampler_dict
 
     @classmethod
@@ -45,36 +98,35 @@ class SamplerData(object):
         if sampler is None:
             return None
         return cls(
-            mag_filter=sampler.get('magFilter'),
-            min_filter=sampler.get('minFilter'),
-            wrap_s=sampler.get('wrapS'),
-            wrap_t=sampler.get('wrapT'),
-            name=sampler.get('name'),
-            extras=sampler.get('extras'),
-            extensions=sampler.get('extensions'),
+            mag_filter=sampler.get("magFilter"),
+            min_filter=sampler.get("minFilter"),
+            wrap_s=sampler.get("wrapS"),
+            wrap_t=sampler.get("wrapT"),
+            name=sampler.get("name"),
+            extras=sampler.get("extras"),
+            extensions=cls.extensions_from_data(sampler.get("extensions")),
         )
 
 
-class TextureData(object):
+class TextureData(BaseGLTFDataClass):
     def __init__(self, sampler=None, source=None, name=None, extras=None, extensions=None):
+        super(TextureData, self).__init__(extras, extensions)
         self.sampler = sampler
         self.source = source
         self.name = name
-        self.extras = extras
-        self.extensions = extensions
 
     def to_data(self, sampler_index_by_key, image_index_by_key):
         texture_dict = {}
         if self.sampler is not None:
-            texture_dict['sampler'] = sampler_index_by_key[self.sampler]
+            texture_dict["sampler"] = sampler_index_by_key[self.sampler]
         if self.source is not None:
-            texture_dict['source'] = image_index_by_key[self.source]
+            texture_dict["source"] = image_index_by_key[self.source]
         if self.name is not None:
-            texture_dict['name'] = self.name
+            texture_dict["name"] = self.name
         if self.extras is not None:
-            texture_dict['extras'] = self.extras
+            texture_dict["extras"] = self.extras
         if self.extensions is not None:
-            texture_dict['extensions'] = self.extensions
+            texture_dict["extensions"] = self.extensions_to_data()
         return texture_dict
 
     @classmethod
@@ -82,29 +134,31 @@ class TextureData(object):
         if texture is None:
             return None
         return cls(
-            sampler=texture.get('sampler'),
-            source=texture.get('source'),
-            name=texture.get('name'),
-            extras=texture.get('extras'),
-            extensions=texture.get('extensions'),
+            sampler=texture.get("sampler"),
+            source=texture.get("source"),
+            name=texture.get("name"),
+            extras=texture.get("extras"),
+            extensions=cls.extensions_from_data(texture.get("extensions")),
         )
 
 
-class TextureInfoData(object):
+class TextureInfoData(BaseGLTFDataClass):
+
+    IS_TEXTURE_INFO_DATA = True  # only needed for ipy in `GLTFContent.check_extensions_texture_recursively`
+
     def __init__(self, index, tex_coord=None, extras=None, extensions=None):
+        super(TextureInfoData, self).__init__(extras, extensions)
         self.index = index
         self.tex_coord = tex_coord
-        self.extras = extras
-        self.extensions = extensions
 
     def to_data(self, texture_index_by_key):
-        texture_info_dict = {'index': texture_index_by_key[self.index]}
+        texture_info_dict = {"index": texture_index_by_key[self.index]}
         if self.tex_coord is not None:
-            texture_info_dict['texCoord'] = self.tex_coord
+            texture_info_dict["texCoord"] = self.tex_coord
         if self.extras is not None:
-            texture_info_dict['extras'] = self.extras
+            texture_info_dict["extras"] = self.extras
         if self.extensions is not None:
-            texture_info_dict['extensions'] = self.extensions
+            texture_info_dict["extensions"] = self.extensions_to_data()
         return texture_info_dict
 
     @classmethod
@@ -112,10 +166,10 @@ class TextureInfoData(object):
         if texture_info is None:
             return None
         return cls(
-            index=texture_info['index'],
-            tex_coord=texture_info.get('texCoord'),
-            extras=texture_info.get('extras'),
-            extensions=texture_info.get('extensions'),
+            index=texture_info["index"],
+            tex_coord=texture_info.get("texCoord"),
+            extras=texture_info.get("extras"),
+            extensions=cls.extensions_from_data(texture_info.get("extensions")),
         )
 
 
@@ -125,9 +179,9 @@ class OcclusionTextureInfoData(TextureInfoData):
         self.strength = strength
 
     def to_data(self, texture_index_by_key):
-        texture_info_dict = super(OcclusionTextureInfoData).to_data(texture_index_by_key)
+        texture_info_dict = super(OcclusionTextureInfoData, self).to_data(texture_index_by_key)
         if self.strength is not None:
-            texture_info_dict['strength'] = self.strength
+            texture_info_dict["strength"] = self.strength
         return texture_info_dict
 
     @classmethod
@@ -135,11 +189,11 @@ class OcclusionTextureInfoData(TextureInfoData):
         if texture_info is None:
             return None
         return cls(
-            index=texture_info['index'],
-            tex_coord=texture_info.get('texCoord'),
-            extras=texture_info.get('extras'),
-            extensions=texture_info.get('extensions'),
-            strength=texture_info.get('strength'),
+            index=texture_info["index"],
+            tex_coord=texture_info.get("texCoord"),
+            extras=texture_info.get("extras"),
+            extensions=cls.extensions_from_data(texture_info.get("extensions")),
+            strength=texture_info.get("strength"),
         )
 
 
@@ -149,9 +203,9 @@ class NormalTextureInfoData(TextureInfoData):
         self.scale = scale
 
     def to_data(self, texture_index_by_key):
-        texture_info_dict = super(NormalTextureInfoData).to_data(texture_index_by_key)
+        texture_info_dict = super(NormalTextureInfoData, self).to_data(texture_index_by_key)
         if self.scale is not None:
-            texture_info_dict['scale'] = self.scale
+            texture_info_dict["scale"] = self.scale
         return texture_info_dict
 
     @classmethod
@@ -159,15 +213,15 @@ class NormalTextureInfoData(TextureInfoData):
         if texture_info is None:
             return None
         return cls(
-            index=texture_info['index'],
-            tex_coord=texture_info.get('texCoord'),
-            extras=texture_info.get('extras'),
-            extensions=texture_info.get('extensions'),
-            scale=texture_info.get('scale'),
+            index=texture_info["index"],
+            tex_coord=texture_info.get("texCoord"),
+            extras=texture_info.get("extras"),
+            extensions=cls.extensions_from_data(texture_info.get("extensions")),
+            scale=texture_info.get("scale"),
         )
 
 
-class PBRMetallicRoughnessData(object):
+class PBRMetallicRoughnessData(BaseGLTFDataClass):
     def __init__(
         self,
         base_color_factor=None,
@@ -178,28 +232,29 @@ class PBRMetallicRoughnessData(object):
         extras=None,
         extensions=None,
     ):
+        super(PBRMetallicRoughnessData, self).__init__(extras, extensions)
         self.base_color_factor = base_color_factor
         self.base_color_texture = base_color_texture
         self.metallic_factor = metallic_factor
         self.roughness_factor = roughness_factor
         self.metallic_roughness_texture = metallic_roughness_texture
-        self.extras = extras
-        self.extensions = extensions
 
     def to_data(self, texture_index_by_key):
         roughness_dict = {}
         if self.base_color_factor is not None:
-            roughness_dict['baseColorFactor'] = self.base_color_factor
+            roughness_dict["baseColorFactor"] = self.base_color_factor
         if self.base_color_texture is not None:
-            roughness_dict['baseColorTexture'] = self.base_color_texture.to_data(texture_index_by_key)
+            roughness_dict["baseColorTexture"] = self.base_color_texture.to_data(texture_index_by_key)
         if self.metallic_factor is not None:
-            roughness_dict['metallicFactor'] = self.metallic_factor
+            roughness_dict["metallicFactor"] = self.metallic_factor
+        if self.roughness_factor is not None:
+            roughness_dict["roughnessFactor"] = self.roughness_factor
         if self.metallic_roughness_texture is not None:
-            roughness_dict['metallicRoughnessTexture'] = self.metallic_roughness_texture.to_data(texture_index_by_key)
+            roughness_dict["metallicRoughnessTexture"] = self.metallic_roughness_texture.to_data(texture_index_by_key)
         if self.extras is not None:
-            roughness_dict['extras'] = self.extras
+            roughness_dict["extras"] = self.extras
         if self.extensions is not None:
-            roughness_dict['extensions'] = self.extensions
+            roughness_dict["extensions"] = self.extensions_to_data()
         return roughness_dict
 
     @classmethod
@@ -207,17 +262,17 @@ class PBRMetallicRoughnessData(object):
         if texture_info is None:
             return None
         return cls(
-            base_color_factor=texture_info.get('baseColorFactor'),
-            base_color_texture=TextureInfoData.from_data(texture_info.get('baseColorTexture')),
-            metallic_factor=texture_info.get('metallicFactor'),
-            roughness_factor=texture_info.get('roughnessFactor'),
-            metallic_roughness_texture=TextureInfoData.from_data(texture_info.get('metallicRoughnessTexture')),
-            extras=texture_info.get('extras'),
-            extensions=texture_info.get('extensions'),
+            base_color_factor=texture_info.get("baseColorFactor"),
+            base_color_texture=TextureInfoData.from_data(texture_info.get("baseColorTexture")),
+            metallic_factor=texture_info.get("metallicFactor"),
+            roughness_factor=texture_info.get("roughnessFactor"),
+            metallic_roughness_texture=TextureInfoData.from_data(texture_info.get("metallicRoughnessTexture")),
+            extras=texture_info.get("extras"),
+            extensions=cls.extensions_from_data(texture_info.get("extensions")),
         )
 
 
-class MaterialData(object):
+class MaterialData(BaseGLTFDataClass):
     def __init__(
         self,
         name=None,
@@ -232,8 +287,8 @@ class MaterialData(object):
         double_sided=None,
         extensions=None,
     ):
+        super(MaterialData, self).__init__(extras, extensions)
         self.name = name
-        self.extras = extras
         self.pbr_metallic_roughness = pbr_metallic_roughness
         self.normal_texture = normal_texture
         self.occlusion_texture = occlusion_texture
@@ -242,32 +297,31 @@ class MaterialData(object):
         self.alpha_mode = alpha_mode
         self.alpha_cutoff = alpha_cutoff
         self.double_sided = double_sided
-        self.extensions = extensions
 
     def to_data(self, texture_index_by_key):
         material_dict = {}
         if self.name is not None:
-            material_dict['name'] = self.name
+            material_dict["name"] = self.name
         if self.extras is not None:
-            material_dict['extras'] = self.extras
+            material_dict["extras"] = self.extras
         if self.pbr_metallic_roughness is not None:
-            material_dict['pbrMetallicRoughness'] = self.pbr_metallic_roughness.to_data(texture_index_by_key)
+            material_dict["pbrMetallicRoughness"] = self.pbr_metallic_roughness.to_data(texture_index_by_key)
         if self.normal_texture is not None:
-            material_dict['normalTexture'] = self.normal_texture.to_data(texture_index_by_key)
+            material_dict["normalTexture"] = self.normal_texture.to_data(texture_index_by_key)
         if self.occlusion_texture is not None:
-            material_dict['materialTexture'] = self.occlusion_texture.to_data(texture_index_by_key)
+            material_dict["materialTexture"] = self.occlusion_texture.to_data(texture_index_by_key)
         if self.emissive_texture is not None:
-            material_dict['emissiveTexture'] = self.emissive_texture.to_data(texture_index_by_key)
+            material_dict["emissiveTexture"] = self.emissive_texture.to_data(texture_index_by_key)
         if self.emissive_factor is not None:
-            material_dict['emissiveFactor'] = self.emissive_factor
+            material_dict["emissiveFactor"] = self.emissive_factor
         if self.alpha_mode is not None:
-            material_dict['alphaMode'] = self.alpha_mode
+            material_dict["alphaMode"] = self.alpha_mode
         if self.alpha_cutoff is not None:
-            material_dict['alphaFactor'] = self.alpha_cutoff
+            material_dict["alphaFactor"] = self.alpha_cutoff
         if self.double_sided is not None:
-            material_dict['doubleSided'] = self.double_sided
+            material_dict["doubleSided"] = self.double_sided
         if self.extensions is not None:
-            material_dict['extensions'] = self.extensions
+            material_dict["extensions"] = self.extensions_to_data(texture_index_by_key=texture_index_by_key)
         return material_dict
 
     @classmethod
@@ -275,41 +329,48 @@ class MaterialData(object):
         if material is None:
             return None
         return cls(
-            name=material.get('name'),
-            extras=material.get('extras'),
-            pbr_metallic_roughness=PBRMetallicRoughnessData.from_data(material.get('pbrMetallicRoughness')),
-            normal_texture=NormalTextureInfoData.from_data(material.get('normalTexture')),
-            occlusion_texture=OcclusionTextureInfoData.from_data(material.get('occlusionTexture')),
-            emissive_texture=TextureInfoData.from_data(material.get('emissiveTexture')),
-            emissive_factor=material.get('emissiveFactor'),
-            alpha_mode=material.get('alphaMode'),
-            alpha_cutoff=material.get('alphaCutoff'),
-            double_sided=material.get('doubleSided'),
-            extensions=material.get('extensions'),
+            name=material.get("name"),
+            extras=material.get("extras"),
+            pbr_metallic_roughness=PBRMetallicRoughnessData.from_data(material.get("pbrMetallicRoughness")),
+            normal_texture=NormalTextureInfoData.from_data(material.get("normalTexture")),
+            occlusion_texture=OcclusionTextureInfoData.from_data(material.get("occlusionTexture")),
+            emissive_texture=TextureInfoData.from_data(material.get("emissiveTexture")),
+            emissive_factor=material.get("emissiveFactor"),
+            alpha_mode=material.get("alphaMode"),
+            alpha_cutoff=material.get("alphaCutoff"),
+            double_sided=material.get("doubleSided"),
+            extensions=cls.extensions_from_data(material.get("extensions")),
         )
 
 
-class CameraData(object):
-    def __init__(self, type_, orthographic=None, perspective=None, name=None, extras=None, extensions=None):
+class CameraData(BaseGLTFDataClass):
+    def __init__(
+        self,
+        type_,
+        orthographic=None,
+        perspective=None,
+        name=None,
+        extras=None,
+        extensions=None,
+    ):
+        super(CameraData, self).__init__(extras, extensions)
         self.type = type_
         self.orthographic = orthographic
         self.perspective = perspective
         self.name = name
-        self.extras = extras
-        self.extensions = extensions
 
     def to_data(self):
-        camera_dict = {'type': self.type}
+        camera_dict = {"type": self.type}
         if self.orthographic is not None:
-            camera_dict['orthographic'] = self.orthographic
+            camera_dict["orthographic"] = self.orthographic
         if self.perspective is not None:
-            camera_dict['perspective'] = self.perspective
+            camera_dict["perspective"] = self.perspective
         if self.name is not None:
-            camera_dict['name'] = self.name
+            camera_dict["name"] = self.name
         if self.extras is not None:
-            camera_dict['extras'] = self.extras
+            camera_dict["extras"] = self.extras
         if self.extensions is not None:
-            camera_dict['extensions'] = self.extensions
+            camera_dict["extensions"] = self.extensions_to_data()
         return camera_dict
 
     @classmethod
@@ -317,34 +378,33 @@ class CameraData(object):
         if camera is None:
             return None
         return cls(
-            type_=camera['type'],
-            orthographic=camera.get('orthographic'),
-            perspective=camera.get('perspective'),
-            name=camera.get('name'),
-            extras=camera.get('extras'),
-            extensions=camera.get('extensions'),
+            type_=camera["type"],
+            orthographic=camera.get("orthographic"),
+            perspective=camera.get("perspective"),
+            name=camera.get("name"),
+            extras=camera.get("extras"),
+            extensions=cls.extensions_from_data(camera.get("extensions")),
         )
 
 
-class AnimationSamplerData(object):
+class AnimationSamplerData(BaseGLTFDataClass):
     def __init__(self, input_, output, interpolation=None, extras=None, extensions=None):
+        super(AnimationSamplerData, self).__init__(extras, extensions)
         self.input = input_
         self.output = output
         self.interpolation = interpolation
-        self.extras = extras
-        self.extensions = extensions
 
     def to_data(self, input_accessor, output_accessor):
         sampler_dict = {
-            'input': input_accessor,
-            'output': output_accessor,
+            "input": input_accessor,
+            "output": output_accessor,
         }
         if self.interpolation is not None:
-            sampler_dict['interpolation'] = self.interpolation
+            sampler_dict["interpolation"] = self.interpolation
         if self.extras is not None:
-            sampler_dict['extras'] = self.extras
+            sampler_dict["extras"] = self.extras
         if self.extensions is not None:
-            sampler_dict['extensions'] = self.extensions
+            sampler_dict["extensions"] = self.extensions_to_data()
         return sampler_dict
 
     @classmethod
@@ -354,29 +414,26 @@ class AnimationSamplerData(object):
         return cls(
             input_=input_,
             output=output,
-            interpolation=sampler.get('interpolation'),
-            extras=sampler.get('extras'),
-            extensions=sampler.get('extensions'),
+            interpolation=sampler.get("interpolation"),
+            extras=sampler.get("extras"),
+            extensions=cls.extensions_from_data(sampler.get("extensions")),
         )
 
 
-class TargetData(object):
+class TargetData(BaseGLTFDataClass):
     def __init__(self, path, node=None, extras=None, extensions=None):
+        super(TargetData, self).__init__(extras, extensions)
         self.path = path
         self.node = node
-        self.extras = extras
-        self.extensions = extensions
 
     def to_data(self, node_index_by_key):
-        target_dict = {
-            'path': self.path
-        }
+        target_dict = {"path": self.path}
         if self.node is not None:
-            target_dict['node'] = node_index_by_key[self.node]
+            target_dict["node"] = node_index_by_key[self.node]
         if self.extras is not None:
-            target_dict['extras'] = self.extras
+            target_dict["extras"] = self.extras
         if self.extensions is not None:
-            target_dict['extensions'] = self.extensions
+            target_dict["extensions"] = self.extensions_to_data()
         return target_dict
 
     @classmethod
@@ -384,29 +441,28 @@ class TargetData(object):
         if target is None:
             return None
         return cls(
-            path=target['path'],
-            node=target.get('node'),
-            extras=target.get('extras'),
-            extensions=target.get('extensions'),
+            path=target["path"],
+            node=target.get("node"),
+            extras=target.get("extras"),
+            extensions=cls.extensions_from_data(target.get("extensions")),
         )
 
 
-class ChannelData(object):
+class ChannelData(BaseGLTFDataClass):
     def __init__(self, sampler, target, extras=None, extensions=None):
+        super(ChannelData, self).__init__(extras, extensions)
         self.sampler = sampler
         self.target = target
-        self.extras = extras
-        self.extensions = extensions
 
     def to_data(self, node_index_by_key, sampler_index_by_key):
         channel_dict = {
-            'sampler': sampler_index_by_key[self.sampler],
-            'target': self.target.to_data(node_index_by_key),
+            "sampler": sampler_index_by_key[self.sampler],
+            "target": self.target.to_data(node_index_by_key),
         }
         if self.extras is not None:
-            channel_dict['extras'] = self.extras
+            channel_dict["extras"] = self.extras
         if self.extensions is not None:
-            channel_dict['extensions'] = self.extensions
+            channel_dict["extensions"] = self.extensions_to_data()
         return channel_dict
 
     @classmethod
@@ -414,38 +470,34 @@ class ChannelData(object):
         if channel is None:
             return None
         return cls(
-            sampler=channel['sampler'],
-            target=TargetData.from_data(channel['target']),
-            extras=channel.get('extras'),
-            extensions=channel.get('extensions'),
+            sampler=channel["sampler"],
+            target=TargetData.from_data(channel["target"]),
+            extras=channel.get("extras"),
+            extensions=cls.extensions_from_data(channel.get("extensions")),
         )
 
 
-class AnimationData(object):
+class AnimationData(BaseGLTFDataClass):
     def __init__(self, channels, samplers_dict, name=None, extras=None, extensions=None):
+        super(AnimationData, self).__init__(extras, extensions)
         self.channels = channels
         self.samplers_dict = samplers_dict
         self.name = name
-        self.extras = extras
-        self.extensions = extensions
 
         self._sampler_index_by_key = None
 
     def to_data(self, samplers_list, node_index_by_key):
-        channels = [
-            channel_data.to_data(node_index_by_key, self._sampler_index_by_key)
-            for channel_data in self.channels
-        ]
+        channels = [channel_data.to_data(node_index_by_key, self._sampler_index_by_key) for channel_data in self.channels]
         animation_dict = {
-            'channels': channels,
-            'samplers': samplers_list,
+            "channels": channels,
+            "samplers": samplers_list,
         }
         if self.name is not None:
-            animation_dict['name'] = self.name
+            animation_dict["name"] = self.name
         if self.extras is not None:
-            animation_dict['extras'] = self.extras
+            animation_dict["extras"] = self.extras
         if self.extensions is not None:
-            animation_dict['extensions'] = self.extensions
+            animation_dict["extensions"] = self.extensions_to_data()
         return animation_dict
 
     def get_sampler_index_by_key(self):
@@ -459,37 +511,41 @@ class AnimationData(object):
         return cls(
             channels=channel_data_list,
             samplers_dict=sampler_dict,
-            name=animation.get('name'),
-            extras=animation.get('extras'),
-            extensions=animation.get('extensions'),
+            name=animation.get("name"),
+            extras=animation.get("extras"),
+            extensions=cls.extensions_from_data(animation.get("extensions")),
         )
 
 
-class SkinData(object):
-    def __init__(self, joints, inverse_bind_matrices=None, skeleton=None, name=None, extras=None, extensions=None):
+class SkinData(BaseGLTFDataClass):
+    def __init__(
+        self,
+        joints,
+        inverse_bind_matrices=None,
+        skeleton=None,
+        name=None,
+        extras=None,
+        extensions=None,
+    ):
+        super(SkinData, self).__init__(extras, extensions)
         self.joints = joints
         self.inverse_bind_matrices = inverse_bind_matrices
         self.skeleton = skeleton
         self.name = name
-        self.extras = extras
-        self.extensions = extensions
 
     def to_data(self, node_index_by_key, accessor_index):
-        node_indices = [
-            node_index_by_key.get(item) for item in self.joints
-            if node_index_by_key.get(item) is not None
-        ]
-        skin_dict = {'joints': node_indices}
+        node_indices = [node_index_by_key.get(item) for item in self.joints if node_index_by_key.get(item) is not None]
+        skin_dict = {"joints": node_indices}
         if self.skeleton is not None:
-            skin_dict['skeleton'] = self.skeleton
+            skin_dict["skeleton"] = self.skeleton
         if self.name is not None:
-            skin_dict['name'] = self.name
+            skin_dict["name"] = self.name
         if self.extras is not None:
-            skin_dict['extras'] = self.extras
+            skin_dict["extras"] = self.extras
         if self.inverse_bind_matrices is not None:
-            skin_dict['inverseBindMatrices'] = accessor_index
+            skin_dict["inverseBindMatrices"] = accessor_index
         if self.extensions is not None:
-            skin_dict['extensions'] = self.extensions
+            skin_dict["extensions"] = self.extensions_from_data()
         return skin_dict
 
     @classmethod
@@ -497,41 +553,47 @@ class SkinData(object):
         if skin is None:
             return None
         return cls(
-            joints=skin['joints'],
+            joints=skin["joints"],
             inverse_bind_matrices=inverse_bind_matrices,
-            skeleton=skin.get('skeleton'),
-            name=skin.get('name'),
-            extras=skin.get('extras'),
-            extensions=skin.get('extensions'),
+            skeleton=skin.get("skeleton"),
+            name=skin.get("name"),
+            extras=skin.get("extras"),
+            extensions=cls.extensions_from_data(skin.get("extensions")),
         )
 
 
-class ImageData(object):
-    def __init__(self, data=None, uri=None, mime_type=None, name=None, extras=None, extensions=None):
+class ImageData(BaseGLTFDataClass):
+    def __init__(
+        self,
+        data=None,
+        uri=None,
+        mime_type=None,
+        name=None,
+        extras=None,
+        extensions=None,
+    ):
+        super(ImageData, self).__init__(extras, extensions)
         self.uri = uri
         self.mime_type = mime_type
         self.name = name
-        self.extras = extras
-        self.extensions = extensions
-
         self.data = data
 
     def to_data(self, uri, buffer_view):
         image_dict = {}
         if self.name is not None:
-            image_dict['name'] = self.name
+            image_dict["name"] = self.name
         if self.extras is not None:
-            image_dict['extras'] = self.extras
+            image_dict["extras"] = self.extras
         if self.mime_type is not None:
-            image_dict['mimeType'] = self.mime_type
+            image_dict["mimeType"] = self.mime_type
         if uri is not None:
-            image_dict['uri'] = uri
+            image_dict["uri"] = uri
         elif buffer_view is not None:
-            image_dict['bufferView'] = buffer_view
+            image_dict["bufferView"] = buffer_view
         elif self.uri is not None:
-            image_dict['uri'] = self.uri
+            image_dict["uri"] = self.uri
         if self.extensions is not None:
-            image_dict['extensions'] = self.extensions
+            image_dict["extensions"] = self.extensions_from_data()
         return image_dict
 
     @classmethod
@@ -539,39 +601,47 @@ class ImageData(object):
         if image is None:
             return None
         return cls(
-            uri=image.get('uri'),
-            mime_type=image.get('mimeType') or mime_type,
-            name=image.get('name'),
-            extras=image.get('extras'),
-            extensions=image.get('extensions'),
+            uri=image.get("uri"),
+            mime_type=image.get("mimeType") or mime_type,
+            name=image.get("name"),
+            extras=image.get("extras"),
+            extensions=cls.extensions_from_data(image.get("extensions")),
             data=data,
         )
 
 
-class PrimitiveData(object):
-    def __init__(self, attributes, indices=None, material=None, mode=None, targets=None, extras=None, extensions=None):
+class PrimitiveData(BaseGLTFDataClass):
+    def __init__(
+        self,
+        attributes,
+        indices=None,
+        material=None,
+        mode=None,
+        targets=None,
+        extras=None,
+        extensions=None,
+    ):
+        super(PrimitiveData, self).__init__(extras, extensions)
         self.attributes = attributes or {}
         self.indices = indices
         self.material = material
         self.mode = mode
         self.targets = targets
-        self.extras = extras
-        self.extensions = extensions
 
     def to_data(self, indices_accessor, attributes_dict, targets_dict, material_index_by_key):
-        primitive_dict = {'indices': indices_accessor}
+        primitive_dict = {"indices": indices_accessor}
         if self.material is not None:
-            primitive_dict['material'] = material_index_by_key[self.material]
+            primitive_dict["material"] = material_index_by_key[self.material]
         if self.mode is not None:
-            primitive_dict['mode'] = self.mode
+            primitive_dict["mode"] = self.mode
         if self.extras is not None:
-            primitive_dict['extras'] = self.extras
+            primitive_dict["extras"] = self.extras
         if attributes_dict:
-            primitive_dict['attributes'] = attributes_dict
+            primitive_dict["attributes"] = attributes_dict
         if targets_dict:
-            primitive_dict['targets'] = targets_dict
+            primitive_dict["targets"] = targets_dict
         if self.extensions is not None:
-            primitive_dict['extensions'] = self.extensions
+            primitive_dict["extensions"] = self.extensions_to_data()
         return primitive_dict
 
     @classmethod
@@ -581,9 +651,9 @@ class PrimitiveData(object):
         return cls(
             attributes=attributes,
             indices=indices,
-            material=primitive.get('material'),
-            mode=primitive.get('mode'),
+            material=primitive.get("material"),
+            mode=primitive.get("mode"),
             targets=target_list,
-            extras=primitive.get('extras'),
-            extensions=primitive.get('extensions'),
+            extras=primitive.get("extras"),
+            extensions=cls.extensions_from_data(primitive.get("extensions")),
         )