+++ src/compas/files/gltf/gltf_content.py
@@ -1,14 +1,14 @@
-from __future__ import print_function
-from __future__ import division
 from __future__ import absolute_import
+from __future__ import division
+from __future__ import print_function
 
+from compas.files.gltf.data_classes import TextureInfoData
 from compas.files.gltf.gltf_mesh import GLTFMesh
 from compas.files.gltf.gltf_node import GLTFNode
 from compas.files.gltf.gltf_scene import GLTFScene
 from compas.files.gltf.helpers import get_weighted_mesh_vertices
-from compas.geometry import transform_points
 from compas.geometry import multiply_matrices
-from compas.utilities import download_file_from_remote
+from compas.geometry import transform_points
 
 
 class GLTFContent(object):
@@ -43,6 +43,7 @@ class GLTFContent(object):
     extensions : object
 
     """
+
     def __init__(self):
         self.scenes = {}
         self.default_scene_key = None
@@ -57,6 +58,7 @@ class GLTFContent(object):
         self.images = {}
         self.extras = None
         self.extensions = None
+        self.extensions_used = None
 
     @property
     def default_or_first_scene(self):
@@ -76,7 +78,7 @@ class GLTFContent(object):
         def visit(key):
             node = self.nodes[key]
             if key in visited_nodes:
-                raise Exception('Nodes do not form a rooted forest.')
+                raise Exception("Nodes do not form a rooted forest.")
             visited_nodes.add(key)
             for child_key in node.children:
                 visit(child_key)
@@ -132,11 +134,7 @@ class GLTFContent(object):
                     animation.channels.remove(channel)
                 else:
                     visited_sampler_keys.append(channel.sampler)
-            animation.samplers_dict = {
-                key: animation.samplers_dict[key]
-                for key in animation.samplers_dict
-                if key in visited_sampler_keys
-            }
+            animation.samplers_dict = {key: animation.samplers_dict[key] for key in animation.samplers_dict if key in visited_sampler_keys}
             if not animation.samplers_dict:
                 del self.animations[animation_key]
 
@@ -158,18 +156,22 @@ class GLTFContent(object):
         self._remove_unvisited(material_visit_log, self.materials)
 
         # walk through existing materials and update textures visit log
+        def check_extensions_texture_recursively(item):
+            # get the extensions that are in the attributes
+            for a in dir(item):
+                if not a.startswith("__") and not callable(getattr(item, a)):
+                    # ipy does not like this one: if isinstance(getattr(item, a), TextureInfoData):
+                    if getattr(getattr(item, a), "IS_TEXTURE_INFO_DATA", False):
+                        texture_visit_log[getattr(item, a).index] = True
+                    # ipy does not like this one: elif isinstance(getattr(item, a), BaseGLTFDataClass):
+                    elif getattr(getattr(item, a), "IS_BASE_GLTF_DATA", False):
+                        check_extensions_texture_recursively(getattr(item, a))
+            if item.extensions is not None:
+                for _, e in item.extensions.items():
+                    check_extensions_texture_recursively(e)
+
         for material in self.materials.values():
-            if material.normal_texture is not None:
-                texture_visit_log[material.normal_texture.index] = True
-            if material.occlusion_texture is not None:
-                texture_visit_log[material.occlusion_texture.index] = True
-            if material.emissive_texture is not None:
-                texture_visit_log[material.emissive_texture.index] = True
-            if material.pbr_metallic_roughness is not None:
-                if material.pbr_metallic_roughness.base_color_texture is not None:
-                    texture_visit_log[material.pbr_metallic_roughness.base_color_texture.index] = True
-                if material.pbr_metallic_roughness.metallic_roughness_texture is not None:
-                    texture_visit_log[material.pbr_metallic_roughness.metallic_roughness_texture.index] = True
+            check_extensions_texture_recursively(material)
 
         # remove unvisited textures
         self._remove_unvisited(texture_visit_log, self.textures)
@@ -226,10 +228,7 @@ class GLTFContent(object):
                 cur = self.nodes[cur_key]
                 for child_key in cur.children:
                     child = self.nodes[child_key]
-                    child.transform = multiply_matrices(
-                        cur.transform,
-                        child.matrix or child.get_matrix_from_trs()
-                    )
+                    child.transform = multiply_matrices(cur.transform, child.matrix or child.get_matrix_from_trs())
                     child.position = transform_points([origin], child.transform)[0]
                     queue.append(child_key)
 
@@ -267,6 +266,95 @@ class GLTFContent(object):
             return mesh_data.vertices
         return get_weighted_mesh_vertices(mesh_data, node.weights)
 
+    def get_node_by_name(self, name):
+        """Returns the node with a specific name.
+
+        Parameters
+        ----------
+        name : str
+            The name of the node
+
+        Returns
+        -------
+        node : :class:`compas.files.GLTFNode` or `None`
+        """
+        for key in self.nodes:
+            if self.nodes[key].name == name:
+                return self.nodes[key]
+        return None
+
+    @classmethod
+    def _get_next_available_key(cls, adict):
+        key = len(adict)
+        while key in adict:
+            key += 1
+        return key
+
+    def add_material(self, material):
+        """Adds a material to the content.
+
+        Parameters
+        ----------
+        material : :class:`compas.files.data_classes.MaterialData`
+            The material to add
+
+        Returns
+        -------
+        int
+        """
+        key = self._get_next_available_key(self.materials)
+        self.materials[key] = material
+        return key
+
+    def add_texture(self, texture):
+        """Adds a texture to the content.
+
+        Parameters
+        ----------
+        texture : :class:`compas.files.data_classes.TextureData`
+            The texture to add
+
+        Returns
+        -------
+        int
+        """
+        key = self._get_next_available_key(self.textures)
+        self.textures[key] = texture
+        return key
+
+    def add_image(self, image):
+        """Adds an image to the content.
+
+        Parameters
+        ----------
+        image : :class:`compas.files.data_classes.ImageData`
+            The image to add
+
+        Returns
+        -------
+        int
+        """
+        key = self._get_next_available_key(self.images)
+        self.images[key] = image
+        return key
+
+    def get_material_index_by_name(self, name):
+        """Returns the index of the material.
+
+        Parameters
+        ----------
+        name : str
+            The name of the material
+
+        Returns
+        -------
+        int or None
+        """
+        for key, material in self.materials.items():
+            if material.name == name:
+                return key
+        return None
+
     def add_scene(self, name=None, extras=None):
         """Adds a scene to the content.
 
@@ -295,7 +383,7 @@ class GLTFContent(object):
         :class:`~compas.files.GLTFNode`
         """
         if scene not in self.scenes.values():
-            raise Exception('Cannot find scene.')
+            raise Exception("Cannot find scene.")
         node = GLTFNode(self, node_name, node_extras)
         scene.children.append(node.key)
         return node
@@ -385,7 +473,7 @@ class GLTFContent(object):
         -------
         tuple
         """
-        positions_dict = {'root': [0, 0, 0]}
+        positions_dict = {"root": [0, 0, 0]}
         edges_list = []
 
         def visit(node, key):
@@ -394,7 +482,7 @@ class GLTFContent(object):
                 edges_list.append((key, child_key))
                 visit(self.nodes[child_key], child_key)
 
-        visit(scene, 'root')
+        visit(scene, "root")
 
         return positions_dict, edges_list
 
@@ -403,30 +491,121 @@ class GLTFContent(object):
 # Main
 # ==============================================================================
 
-if __name__ == '__main__':
+if __name__ == "__main__":
 
     import os
+    import urllib
+
     import compas
-
     from compas.datastructures import Mesh
+    from compas.files.gltf.data_classes import ImageData
+    from compas.files.gltf.data_classes import MaterialData
+    from compas.files.gltf.data_classes import MineType
+    from compas.files.gltf.data_classes import PBRMetallicRoughnessData
+    from compas.files.gltf.data_classes import TextureData
+    from compas.files.gltf.extensions import KHR_materials_pbrSpecularGlossiness
+    from compas.files.gltf.extensions import KHR_Texture_Transform
     from compas.files.gltf.gltf import GLTF
+    from compas.geometry import Box
+    from compas.geometry import Frame
+    from compas.utilities import download_file_from_remote
 
-    source = 'https://raw.githubusercontent.com/ros-industrial/abb/kinetic-devel/abb_irb6600_support/meshes/irb6640/visual/link_1.stl'
-    stl_filepath = os.path.join(compas.APPDATA, 'data', 'meshes', 'ros', 'link_1.stl')
+    dirname = os.path.join(compas.APPDATA, "data", "gltfs")
+    gltf_filepath = os.path.join(dirname, "compas.gltf")
 
-    download_file_from_remote(source, stl_filepath, overwrite=False)
+    image_uri = "compas_icon_white.png"
+    image_file = os.path.join(dirname, image_uri)
+    try:
+        download_file_from_remote("https://compas.dev/images/compas_icon_white.png", image_file)
+    except urllib.error.HTTPError:
+        pass
 
-    gltf_filepath = os.path.join(compas.APPDATA, 'data', 'gltfs', 'double_link_1.gltf')
-
-    mesh = Mesh.from_stl(stl_filepath)
     cnt = GLTFContent()
     scene = cnt.add_scene()
-    node_1 = scene.add_child(node_name='Node1')
-    mesh_data = node_1.add_mesh(mesh)
-    node_2 = node_1.add_child(child_name='Node2')
-    node_2.translation = [0, 0, 5]
-    node_2.add_mesh(mesh_data.key)
+
+    # image's uri should be the relative path to the image from the filepath given at the time of export,
+    # so if the image will sit in the same directory as the resultant gltf, the uri is just the name of the file.
+    # it's the exporter's job to manage how things are stored in the buffer, and it would only store image data
+    # in the buffer if it is exporting as glb. otherwise the uri will just stay the relative path to the image
+    # and the gltf only makes sense when bundled with these external files.
+    image_data = ImageData(
+        name=image_uri,
+        mime_type=MineType.PNG,
+        uri=image_file,
+    )
+    image_idx = cnt.add_image(image_data)
+    # TextureData.source takes the key of the ImageData that it should use as 'source'
+    texture = TextureData(source=image_idx)
+    texture_idx = cnt.add_texture(texture)
+
+    texture = TextureData(source=image_idx)
+    texture_idx2 = cnt.add_texture(texture)
+
+    material = MaterialData()
+    material.name = "Texture"
+    material.pbr_metallic_roughness = PBRMetallicRoughnessData()
+    material.pbr_metallic_roughness.metallic_factor = 0.0
+    material.pbr_metallic_roughness.base_color_texture = TextureInfoData(index=texture_idx)
+    material_key = cnt.add_material(material)
+
+    # add extension
+    pbr_specular_glossiness = KHR_materials_pbrSpecularGlossiness()
+    pbr_specular_glossiness.diffuse_factor = [0.980392158, 0.980392158, 0.980392158, 1.0]
+    pbr_specular_glossiness.specular_factor = [0.0, 0.0, 0.0]
+    pbr_specular_glossiness.glossiness_factor = 0.0
+    texture_transform = KHR_Texture_Transform()
+    texture_transform.rotation = 0.0
+    texture_transform.scale = [2.0, 2.0]
+    # same here, TextureInfoData uses the key of the TextureData
+    pbr_specular_glossiness.diffuse_texture = TextureInfoData(texture_idx2)
+    pbr_specular_glossiness.diffuse_texture.add_extension(texture_transform)
+    material.add_extension(pbr_specular_glossiness)
+
+    # add box
+    box = Box(Frame.worldXY(), 1, 1, 1)
+    mesh = Mesh.from_shape(box)
+    mesh.quads_to_triangles()
+
+    node = scene.add_child()
+    mesh_data = node.add_mesh(mesh)
+    normals = [mesh.vertex_normal(k) for k in mesh.vertices()]
+
+    texcoord_0 = [(0, 0) for _ in mesh.vertices()]
+
+    """
+    for fkey in mesh.faces():
+        vkeys = mesh.face_vertices(fkey)
+        plane = mesh.face_plane(fkey)
+        frame = Frame.from_plane(plane)
+        coords = mesh.face_coordinates(fkey)
+        for vkey, xyz in zip(vkeys, coords):
+            u, v, _ = frame.to_local_coordinates(Point(*xyz))
+            texcoord_0[vkey] = (u, v)  # not ideal, gets overwritten
+    """
+
+    # here is the tricky part... for this material to be valid and applied to this mesh,
+    # each of the primitives must have within the attribute `attributes` a key of the form `TEXCOORD_{some integer}`.
+    # the value of this thing should be a list of pairs of floats representing the UV texture coordinates for each vertex.
+    # if `{some integer}` is 0 then there's nothing else to do.  but if a primitive has multiple `TEXTCOORD_{some integer}`s,
+    # then the various `TextureInfoData.tex_coord` associated to this material have to be updated with the appropriate `{some integer}`.
+
+    # would work better if each vertex could have 4 different texture coordinates
+    texcoord_0 = [
+        (0.0, 1.0),
+        (0.0, 0.0),
+        (1.0, 0.0),
+        (1.0, 1.0),
+        (0.0, 0.0),
+        (1.0, 0.0),
+        (1.0, 1.0),
+        (0.0, 1.0),
+    ]
+
+    pd = node.mesh_data.primitive_data_list[0]
+    pd.material = material_key
+    pd.attributes["TEXCOORD_0"] = texcoord_0
+    pd.attributes["NORMAL"] = normals
 
     gltf = GLTF(gltf_filepath)
     gltf.content = cnt
-    gltf.export(embed_data=True)
+    gltf.export(embed_data=False)